+++ api/models/Calendar.js
@@ -27,8 +27,10 @@ module.exports = {
         },
 
         active: {
-            type: 'boolean',
-            defaultsTo: true
+            type: 'number',
+            min: -1,
+            max: 1,
+            defaultsTo: 1
         },
 
         title: {
@@ -519,7 +521,6 @@ module.exports = {
                                     if (key === `end` && moment(theEvent[key]).isSame(moment(criteria[key])))
                                         continue;
                                     needsUpdate = true;
-                                    sails.log.warn(`Needs update: ${key} WHERE ${criteria[key]} !== ${theEvent[key]} .`);
                                     break;
                                 }
                             }
@@ -590,18 +591,91 @@ module.exports = {
                     }
 
                     // Update entries in the calendar which passed their end time
-                    var destroyed = await Calendar.update({active: true, end: {"<=": moment().toISOString(true)}}, {active: false})
+                    var destroyed = await Calendar.update({active: 1, end: {"<=": moment().toISOString(true)}}, {active: 0})
                             .tolerate((err) => {
                             })
                             .fetch();
 
-                    // Destroy entries in the calendar which no longer exist on Google Calendar
                     if (events.length > 0)
                     {
-                        await Calendar.destroy({unique: {'nin': eventIds}})
+                        // Destroy events which ended before midnight of today's day.
+                        await Calendar.destroy({end: {'<': moment().startOf('day').toISOString(true)}})
                                 .tolerate((err) => {
                                 })
                                 .fetch();
+
+                        // Entries no longer in Google Calendar should be updated to active = -1 to indicate they were canceled.
+                        var cancelled = await Calendar.update({unique: {'nin': eventIds}, active: 1}, {active: -1})
+                                .tolerate((err) => {
+                                })
+                                .fetch();
+
+                        // Send out cancellation notifications for cancelled shows. Also add a log for cancelled shows.
+                        if (cancelled.length > 1)
+                        {
+                            cancelled.map((cEvent) => {
+                                (async () => {
+                                    var dj = cEvent.title;
+                                    if (dj.includes(" - ") && dj.includes(": "))
+                                    {
+                                        dj = dj.split(" - ")[0];
+                                        dj = dj.substring(dj.indexOf(": ") + 2);
+                                    } else {
+                                        dj = null;
+                                    }
+                                    if (dj !== null)
+                                        dj = await Djs.findOrCreate({name: dj}, {name: dj, lastSeen: moment("2002-01-01").toISOString(true)});
+                                    var attendance = await Attendance.create({unique: cEvent.unique, dj: dj !== null && typeof dj.ID !== 'undefined' ? dj.ID : null, event: `${cEvent.title} {CANCELLED via Google Calendar}`, scheduledStart: moment(cEvent.start).toISOString(true), scheduledEnd: moment(cEvent.end).toISOString(true)});
+                                    if (cEvent.title.startsWith("Show: "))
+                                    {
+                                        var temp = cEvent.title.replace("Show: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Show: `, temp, `Live Show`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                        await Logs.create({attendanceID: attendance.ID, logtype: 'cancellation', loglevel: 'warning', logsubtype: temp, event: `Show was cancelled via Google Calendar!<br />Show: ${temp}<br />Scheduled time: ${moment(cEvent.start).format("LLL")} - ${moment(cEvent.end).format("LT")}`, createdAt: moment().toISOString(true)}).fetch()
+                                                .tolerate((err) => {
+                                                    sails.log.error(err);
+                                                });
+                                    }
+                                    if (cEvent.title.startsWith("Remote: "))
+                                    {
+                                        var temp = cEvent.title.replace("Remote: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Remote: `, temp, `Remote Broadcast`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                        await Logs.create({attendanceID: attendance.ID, logtype: 'cancellation', loglevel: 'warning', logsubtype: temp, event: `Remote broadcast was cancelled via Google Calendar!<br />Remote: ${temp}<br />Scheduled time: ${moment(cEvent.start).format("LLL")} - ${moment(cEvent.end).format("LT")}`, createdAt: moment().toISOString(true)}).fetch()
+                                                .tolerate((err) => {
+                                                    sails.log.error(err);
+                                                });
+                                    }
+                                    if (cEvent.title.startsWith("Sports: "))
+                                    {
+                                        var temp = cEvent.title.replace("Sports: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Sports: `, temp, `Sports Broadcast`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                        await Logs.create({attendanceID: attendance.ID, logtype: 'cancellation', loglevel: 'warning', logsubtype: temp, event: `Sports broadcast was cancelled via Google Calendar!<br />Sports: ${temp}<br />Scheduled time: ${moment(cEvent.start).format("LLL")} - ${moment(cEvent.end).format("LT")}`, createdAt: moment().toISOString(true)}).fetch()
+                                                .tolerate((err) => {
+                                                    sails.log.error(err);
+                                                });
+                                    }
+                                    if (cEvent.title.startsWith("Prerecord: "))
+                                    {
+                                        var temp = cEvent.title.replace("Prerecord: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Prerecord: `, temp, `Prerecorded Show`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                        await Logs.create({attendanceID: attendance.ID, logtype: 'cancellation', loglevel: 'warning', logsubtype: temp, event: `Prerecorded show was cancelled via Google Calendar!<br />Prerecord: ${temp}<br />Scheduled time: ${moment(cEvent.start).format("LLL")} - ${moment(cEvent.end).format("LT")}`, createdAt: moment().toISOString(true)}).fetch()
+                                                .tolerate((err) => {
+                                                    sails.log.error(err);
+                                                });
+                                    }
+                                    // We don't care about logging cancelled genres nor playlists, but we still want to send notifications out.
+                                    if (cEvent.title.startsWith("Genre: "))
+                                    {
+                                        var temp = cEvent.title.replace("Genre: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Genre: `, temp, `Genre`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                    }
+                                    if (cEvent.title.startsWith("Playlist: "))
+                                    {
+                                        var temp = cEvent.title.replace("Playlist: ", "");
+                                        await sails.helpers.onesignal.sendEvent(`Playlist: `, temp, `Playlist`, cEvent.unique, moment(cEvent.start).format("LLL"));
+                                    }
+                                });
+                            });
+                        }
 
                     }
 
@@ -622,7 +696,7 @@ module.exports = {
                                             dj = null;
                                         }
                                         if (dj !== null)
-                                            dj = await Djs.findOne({name: dj});
+                                            dj = await Djs.findOrCreate({name: dj}, {name: dj, lastSeen: moment("2002-01-01").toISOString(true)});
                                         Attendance.findOrCreate({unique: event.unique}, {unique: event.unique, dj: dj !== null && typeof dj.ID !== 'undefined' ? dj.ID : null, event: event.title, scheduledStart: moment(event.start).toISOString(true), scheduledEnd: moment(event.end).toISOString(true)})
                                                 .exec(async(err, record, wasCreated) => {
                                                     if (err)